requirements.txt
# Core dependencies
requests
python-dotenv
praw
openai>=1.0.0

# Optional (used in subprocess execution and date handling)
# These are part of Python standard library but listed for clarity
# subprocess
# datetime
# json
# time
# os
# re
# urllib
